Temporary save whilst working on NotificationBroadcaster.
